extractor/requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
databricks-sql-connector==3.3.0
et-xmlfile==1.1.0
greenlet==3.0.3
idna==3.7
lz4==4.3.3
numpy==1.26.4
oauthlib==3.2.2
openpyxl==3.1.5
pandas==2.1.4
pyarrow==16.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
SQLAlchemy==2.0.32
thrift==0.20.0
tzdata==2024.1
urllib3==2.2.2
databricks-sql-connector[sqlalchemy]